* Fixing another ambiguity caused by `using namespace`.  * Tests for chunk pool allocator.  * Adjusted allocator tests to new assertion interface, renamed tests for better consistency.
git-svn-id: http://svn.seqan.de/seqan/trunk@9094 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/basic/basic_allocator_chunkpool.h
@@ -261,7 +261,7 @@ deallocate(Allocator<ChunkPool<SIZE, MAX_COUNT, TParentAllocator> > & me,
 
 	typedef Allocator<ChunkPool<SIZE, MAX_COUNT, TParentAllocator> > TAllocator;
 
-	if ((sizeof(TValue) != SIZE) || (count > MAX_COUNT))
+	if ((sizeof(TValue) != SIZE) || (static_cast<size_t>(count) > MAX_COUNT))
 	{//no blocking
 		return deallocate(parentAllocator(me), data, count, tag_);
 	}
